re-ordering code for Gating functions to be added
